Add VotingResults method to display vote percentages
The code now prints a blank line and the message "The results are in:" followed by the counts of "Yes" and "No" votes after an invalid vote is detected. A new method `VotingResults` has been added, which calculates and prints the percentage of "Yes" and "No" votes. The `VotingResults` method is called after printing the raw vote counts.